% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gcdg_lean}
\alias{gcdg_lean}
\title{This is data to be included in my package}
\format{
An object of class \code{lean} (inherits from \code{list}) of length 3.
}
\usage{
gcdg_lean
}
\description{
This is data to be included in my package
}
\references{
\url{ddata package}
}
\author{
stef van buuren
}
\keyword{datasets}
