% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{itemtableVWO}
\alias{itemtableVWO}
\title{itemtable van Wiechen}
\format{
An object of class \code{data.frame} with 89 rows and 11 columns.
}
\usage{
itemtableVWO
}
\description{
Created by \code{data-raw/R/itemtableVWO.R}
}
\references{
\url{data-raw/R/itemtableVWO.R}
}
\author{
project-team
}
\keyword{datasets}
