% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{jgz_vwo}
\alias{jgz_vwo}
\title{This is data to be included in my package}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 23097 rows and 108 columns.
}
\usage{
jgz_vwo
}
\description{
This is data to be included in my package
}
\references{
\url{dscoreddi package}
}
\author{
iris eekhout
}
\keyword{datasets}
