% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_child_select.R
\name{mod_child_select_server}
\alias{mod_child_select_server}
\title{Child-selection Shiny module server}
\usage{
mod_child_select_server(id, feedback_data)
}
\arguments{
\item{id}{Module id.}

\item{feedback_data}{A \code{dfeedback_data} object or reactive returning one.}
}
\value{
A reactive containing the selected child id.
}
\description{
Child-selection Shiny module server
}
