% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_child_select.R
\name{mod_child_select_ui}
\alias{mod_child_select_ui}
\title{Child-selection Shiny module UI}
\usage{
mod_child_select_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
Child-selection Shiny module UI
}
