% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_growth_basic.R
\name{mod_growth_basic_server}
\alias{mod_growth_basic_server}
\title{Basic growth-chart module server}
\usage{
mod_growth_basic_server(
  id,
  feedback_data,
  child_id,
  reference_key = NULL,
  reference_population = NULL
)
}
\arguments{
\item{id}{Module id.}

\item{feedback_data}{A \code{dfeedback_data} object or reactive returning one.}

\item{child_id}{Selected child identifier or reactive returning one.}

\item{reference_key}{Optional D-score key for the reference data. If \code{NULL},
the key in \code{feedback_data$settings$key} is used.}

\item{reference_population}{Optional reference population. If \code{NULL},
the population in \code{feedback_data$settings$population} is used.}
}
\value{
Reactive Plotly object.
}
\description{
Basic growth-chart module server
}
