% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_growth_basic.R
\name{mod_growth_basic_ui}
\alias{mod_growth_basic_ui}
\title{Basic growth-chart module UI}
\usage{
mod_growth_basic_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
Shiny UI elements.
}
\description{
Basic growth-chart module UI
}
