% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_growth_domains.R
\name{mod_growth_domains_server}
\alias{mod_growth_domains_server}
\title{Domain growth-chart module server}
\usage{
mod_growth_domains_server(id, feedback_data, child_id)
}
\description{
Domain growth-chart module server
}
