% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_growth_domains.R
\name{mod_growth_domains_ui}
\alias{mod_growth_domains_ui}
\title{Domain growth-chart module UI}
\usage{
mod_growth_domains_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
Domain growth-chart module UI
}
