% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_growth_item_drilldown.R
\name{mod_growth_item_drilldown_server}
\alias{mod_growth_item_drilldown_server}
\title{D-score growth curve with clickable item details}
\usage{
mod_growth_item_drilldown_server(id, feedback_data, child_id)
}
\arguments{
\item{id}{Module id.}

\item{feedback_data}{A \code{dfeedback_data} object or reactive returning one.}

\item{child_id}{Selected child identifier or reactive returning one.}
}
\value{
A reactive containing the selected contact moment.
}
\description{
D-score growth curve with clickable item details
}
