% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_growth_item_drilldown.R
\name{mod_growth_item_drilldown_ui}
\alias{mod_growth_item_drilldown_ui}
\title{D-score growth curve with clickable item details}
\usage{
mod_growth_item_drilldown_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
Shiny UI elements.
}
\description{
D-score growth curve with clickable item details
}
