% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_growth_item_overlay.R
\name{mod_growth_item_overlay_server}
\alias{mod_growth_item_overlay_server}
\title{D-score curve with item-level observations around each score}
\usage{
mod_growth_item_overlay_server(id, feedback_data, child_id)
}
\arguments{
\item{id}{Module id.}

\item{feedback_data}{A \code{dfeedback_data} object or reactive returning one.}

\item{child_id}{Selected child identifier or reactive returning one.}
}
\value{
Reactive Plotly object.
}
\description{
D-score curve with item-level observations around each score
}
