% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_growth_item_overlay.R
\name{mod_growth_item_overlay_ui}
\alias{mod_growth_item_overlay_ui}
\title{D-score curve with item-level observations around each score}
\usage{
mod_growth_item_overlay_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
Shiny UI elements.
}
\description{
D-score curve with item-level observations around each score
}
