% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_item_timeline.R
\name{mod_item_timeline_server}
\alias{mod_item_timeline_server}
\title{Item timeline module server}
\usage{
mod_item_timeline_server(id, feedback_data, child_id)
}
\arguments{
\item{id}{Module id.}

\item{feedback_data}{A \code{dfeedback_data} object or reactive returning one.}

\item{child_id}{Selected child identifier or reactive returning one.}
}
\value{
Reactive Plotly object.
}
\description{
Item timeline module server
}
