% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_item_timeline.R
\name{mod_item_timeline_ui}
\alias{mod_item_timeline_ui}
\title{Item timeline module UI}
\usage{
mod_item_timeline_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
Shiny UI elements.
}
\description{
Item timeline module UI
}
