% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_plot_data.R
\name{percentile_plot_data}
\alias{percentile_plot_data}
\title{Percentile plot data}
\usage{
percentile_plot_data(itembank, reference, age, refperc, suggest)
}
\arguments{
\item{itembank}{itembank used to obtain suggested items given age an percentile.}

\item{reference}{reference for d-score given age}

\item{age}{age of the child (dynamic with use)}

\item{refperc}{percentile used to select suggetsed items (dynamic with use)}

\item{suggest}{number of suggested items (dynamic with use)}
}
\value{
data.frame used for plot
}
\description{
Create the input data to plot the percentile references for each Van Wiechen
milestone for P10, P50 and P90, per domain.
}
