% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vwc.R
\name{plot_vwc}
\alias{plot_vwc}
\title{Van Wiechen Continue plot}
\usage{
plot_vwc(
  data = dscoreddi::prefdat,
  domein = "Grove motoriek",
  age = NULL,
  selected = NULL,
  vw_history = NULL
)
}
\arguments{
\item{data}{data.frame with columsn as in \code{dscoreddi::prefdat}, c("A10", "A50", "A90", "domein", "labelNL", "nr", "labelNLnr", "sideA", "sideB").}

\item{domein}{character vector for domein filter, any of c("Fijne motoriek", "Communicatie","Grove motoriek")}

\item{age}{age in decimal years}

\item{selected}{character vector with item names of selected items.}

\item{vw_history}{data.frame with history of vw responses column c("age", "item", "value")}
}
\value{
plot_ly object
}
\description{
Draw plotly object of domain specific plot with VWC items.
}
\examples{
plot_vwc()
}
