% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{prefdat}
\alias{prefdat}
\title{This is data to be included in my package}
\format{
An object of class \code{data.frame} with 81 rows and 21 columns.
}
\usage{
prefdat
}
\description{
This is data to be included in my package
}
\references{
\url{dscoreddi package}
}
\author{
iris eekhout
}
\keyword{datasets}
