% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_vwc.R
\name{select_vwc}
\alias{select_vwc}
\title{Select Van Wiechen Continue}
\usage{
select_vwc(
  data = dscoreddi::prefdat,
  itembank = dscoreddi::itembank_vwc,
  age = NULL,
  refperc = 90,
  nsuggest = 6,
  vw_history = NULL,
  passed_items = NULL
)
}
\arguments{
\item{data}{data.frame to select items from with at least column item}

\item{itembank}{data.frame item and tau}

\item{age}{age in decimal years}

\item{refperc}{percentile to base item selection on (given age)}

\item{nsuggest}{number of items to select, default = 6}

\item{vw_history}{data.frame with historical vw data with colums c("age", "item", "value")}

\item{passed_items}{character vector with names of passed_items to exclude. Only used when vw_history not available.}
}
\value{
character vector
}
\description{
Select the suggested items to be shown in the VWC plot.
}
\examples{
select_vwc(age = 2)
}
