% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{smoccmodel}
\alias{smoccmodel}
\title{This is data to be included in my package}
\format{
An object of class \code{list} of length 4.
}
\usage{
smoccmodel
}
\description{
This is data to be included in my package
}
\references{
\url{ddata package}
}
\author{
stef van buuren
}
\keyword{datasets}
