% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwc_percentiles.R
\name{vwc_percentiles}
\alias{vwc_percentiles}
\title{VWC percentiles}
\usage{
vwc_percentiles(
  items,
  p = c(2, 10, 50, 90, 98),
  key = "dutch",
  population = "dutch",
  reference = filter(dscoreddi::expanded_reference, population == "dutch" & key ==
    "dutch")
)
}
\arguments{
\item{items}{character vector with items names for selected items}

\item{p}{percentiles default: \code{p = c(2, 10, 50, 90, 98)}}

\item{key}{D-score model key}

\item{population}{Reference population}

\item{reference}{data.frame with reference standard, if \code{reference = NULL} the
\code{key} and \code{population} arguments are used.}
}
\value{
data.frame with D-score percentiles and Age percentiles, age in months.
}
\description{
VWC percentiles
}
\examples{
vwc_percentiles(select_vwc(age = 2))
}
