% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwo_domein_plot.R
\name{vwo_domein_plot}
\alias{vwo_domein_plot}
\title{Van Wiechen continue plot for domain}
\usage{
vwo_domein_plot(data, domein, side, age)
}
\arguments{
\item{data}{data.frame with all information to use in plot, created with
\code{percentile_plot_data} function.}

\item{domein}{domain to be plotted}

\item{side}{character vector "sideA" or "sideB" indicating the side of VWO
schema}

\item{age}{age of the child that is measured in months}
}
\value{
ggplot
}
\description{
Van Wiechen continue plot for domain
}
\examples{
library(dplyr)
itembank <- dscore::builtin_itembank \%>\%
filter(key == "dutch") \%>\%
bind_rows(data.frame(item = "ddifmd028", tau = 72),
          data.frame(item = "ddicmm051", tau = 74),
          data.frame(item = "ddigmd075", tau = 72)
          )
expanded_reference <- dscoreddi::expand_reference(population = "dutch")
pref <- percentile_plot_data(itembank = itembank,
reference = expanded_reference, age = 6, refperc = 90, suggest = 6)
p1 <- vwo_domein_plot(data = pref, domein = "Communicatie",  side = "sideA", age = 6)
p1
}
