% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_reference.R
\name{expand_reference}
\alias{expand_reference}
\title{Expand reference}
\usage{
expand_reference(population = "gcdg")
}
\arguments{
\item{population}{A string describing the population. Currently supported
are "dutch" and "gcdg" (default).}
}
\value{

}
\description{
Expand reference obtained with \code{dscore::get_reference} with a count model
(based on dmetric/expand_referenced.Rmd)
44.35 - 1.8 * t + 28.47 * log(t + 0.25)
}
\examples{
expand_reference(population = "dutch")
}
