% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{expanded_reference}
\alias{expanded_reference}
\title{This is data to be included in my package}
\format{
An object of class \code{data.frame} with 514 rows and 4 columns.
}
\usage{
expanded_reference
}
\description{
This is data to be included in my package
}
\references{
\url{dscoreddi package}
}
\author{
iris eekhout
}
\keyword{datasets}
