% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_reference.R
\name{expanded_reference}
\alias{expanded_reference}
\title{Expand reference}
\usage{
expanded_reference(
  reference = dscore::get_reference(population = "gcdg"),
  population = reference$pop[1]
)
}
\arguments{
\item{reference}{A data.frame with the same structure as builtin_references.
The default is to use dscore::builtin_references.}

\item{population}{A string describing the population. Currently supported
are "dutch" and "gcdg" (default).}
}
\value{

}
\description{
Expand reference with a count model (based on dmetric/expand_referenced.Rmd)
44.35 - 1.8 * t + 28.47 * log(t + 0.25)
}
