% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{itembank_vwc}
\alias{itembank_vwc}
\title{This is data to be included in my package}
\format{
An object of class \code{data.frame} with 81 rows and 8 columns.
}
\usage{
itembank_vwc
}
\description{
This is data to be included in my package
}
\references{
\url{dscoreddi package}
}
\author{
iris eekhout
}
\keyword{datasets}
