% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_vwo_gsed.R
\name{rename_vwo_gsed}
\alias{rename_vwo_gsed}
\title{Rename VWO items to gsed naming scheme.}
\usage{
rename_vwo_gsed(x, copy = TRUE, version = "2005")
}
\arguments{
\item{x}{character vector with the ID.VWO2005 item name, starting with "v".}

\item{copy}{A logical indicating whether any unmatches names should
be copied (\code{copy = TRUE}) or set to an empty string.}

\item{version}{default = "2005"; version 1996 also possible.}
}
\value{
A character vector of length \code{length(x)} with VWO
item names replaced by gsed item name.
}
\description{
Rename VWO2005 item names to names in GSED format.
}
\details{
The gsed-naming convention is as follows. Position 1-3 codes the
instrument, position 4-5 codes the domain, position 6 codes
direct/caregiver/message, positions 7-9 is a item sequence number.
The object itemtableVWO is used as a renaming table.
}
\examples{
rename_vwo_gsed("v1")
rename_vwo_gsed(c("v1", "v40"))
rename_vwo_gsed(c("v55", "v55a", "v55c"))
}
